+++ src/types/fields.types.ts
@@ -1,4 +1,11 @@
-export type SELECT_FIELD_VALUE = {
+export type SelectFieldValue = {
   label: string | number
   value: string | number
+}
+
+export type InputDropdownPickOption = {
+  id: number
+  text: string
+  image?: string | null
+  followersCount?: number
 }